+++ events_protocol/core/logging/mixins/loggable.py
@@ -1,0 +1,5 @@
+from events_protocol.core.logging import JsonLogger
+
+
+class LoggableMixin:
+    logger = JsonLogger()