+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.1.0] - 13-02-20
+**Author**
+
+Romulo Rosa - romulo.rosa@guiabolso.com.br
+
+### Added
+- Add server module (Handler, Discovery and Register)
+
+### Changes
+- Use pydantic as type checker
+
 ## [0.0.1] - 13-02-20
 **Author**